csharp-mvc: version history

+++ web/documentserver-example/csharp-mvc/WebEditor.ashx.cs
@@ -95,6 +95,7 @@ namespace OnlineEditorsExampleMVC
 
                 var savedFileName = DocManagerHelper.StoragePath(fileName);
                 httpPostedFile.SaveAs(savedFileName);
+                DocManagerHelper.CreateMeta(fileName, context.Request.Cookies["uid"]?.Value, context.Request.Cookies["uname"]?.Value);
 
                 context.Response.Write("{ \"filename\": \"" + fileName + "\"}");
             }
@@ -148,8 +149,9 @@ namespace OnlineEditorsExampleMVC
                         }
                     }
 
-                    File.Delete(DocManagerHelper.StoragePath(fileName));
+                    Remove(fileName);
                     fileName = correctName;
+                    DocManagerHelper.CreateMeta(fileName, context.Request.Cookies["uid"]?.Value, context.Request.Cookies["uname"]?.Value);
                 }
 
                 context.Response.Write("{ \"filename\" : \"" + fileName + "\"}");
@@ -173,8 +175,6 @@ namespace OnlineEditorsExampleMVC
         {
             var userAddress = context.Request["userAddress"];
             var fileName = context.Request["fileName"];
-
-            var storagePath = DocManagerHelper.StoragePath(fileName, userAddress);
 
             string body;
             try
@@ -219,26 +219,32 @@ namespace OnlineEditorsExampleMVC
                 case TrackerStatus.Corrupted:
                     var downloadUri = (string) fileData["url"];
 
-                    var req = (HttpWebRequest) WebRequest.Create(downloadUri);
-
                     var saved = 1;
                     try
                     {
-                        using (var stream = req.GetResponse().GetResponseStream())
-                        {
-                            if (stream == null) throw new Exception("stream is null");
-                            const int bufferSize = 4096;
+                        var storagePath = DocManagerHelper.StoragePath(fileName, userAddress);
+                        var histDir = DocManagerHelper.HistoryDir(storagePath);
+                        var versionDir = DocManagerHelper.VersionDir(histDir, DocManagerHelper.GetFileVersion(histDir) + 1);
 
-                            using (var fs = File.Open(storagePath, FileMode.Create))
-                            {
-                                var buffer = new byte[bufferSize];
-                                int readed;
-                                while ((readed = stream.Read(buffer, 0, bufferSize)) != 0)
-                                {
-                                    fs.Write(buffer, 0, readed);
-                                }
-                            }
+                        if (!Directory.Exists(versionDir)) Directory.CreateDirectory(versionDir);
+
+                        File.Copy(storagePath, Path.Combine(versionDir, "prev" + Path.GetExtension(fileName)));
+
+                        DownloadToFile(downloadUri, DocManagerHelper.StoragePath(fileName, userAddress));
+                        DownloadToFile((string)fileData["changesurl"], Path.Combine(versionDir, "diff.zip"));
+
+                        var hist = fileData.ContainsKey("changeshistory") ? (string)fileData["changeshistory"] : null;
+                        if (string.IsNullOrEmpty(hist) && fileData.ContainsKey("history"))
+                        {
+                            hist = jss.Serialize(fileData["history"]);
                         }
+
+                        if (!string.IsNullOrEmpty(hist))
+                        {
+                            File.WriteAllText(Path.Combine(versionDir, "changes.json"), hist);
+                        }
+
+                        File.WriteAllText(Path.Combine(versionDir, "key.txt"), (string)fileData["key"]);
                     }
                     catch (Exception)
                     {
@@ -256,16 +262,45 @@ namespace OnlineEditorsExampleMVC
             try
             {
                 var fileName = context.Request["fileName"];
-                var directory = HttpRuntime.AppDomainAppPath + WebConfigurationManager.AppSettings["storage-path"] + DocManagerHelper.CurUserHostAddress(null) + "\\";
-                var path = Path.Combine(directory, fileName);
-
-                File.Delete(path);
+                Remove(fileName);
 
                 context.Response.Write("{ \"success\": true }");
             }
             catch (Exception e)
             {
                 context.Response.Write("{ \"error\": \"" + e.Message + "\"}");
+            }
+        }
+
+        private static void Remove(string fileName)
+        {
+            var path = DocManagerHelper.StoragePath(fileName, null);
+            var histDir = DocManagerHelper.HistoryDir(path);
+
+            if (File.Exists(path)) File.Delete(path);
+            if (Directory.Exists(histDir)) Directory.Delete(histDir, true);
+        }
+
+        private static void DownloadToFile(string url, string path)
+        {
+            if (string.IsNullOrEmpty(url)) throw new ArgumentException("url");
+            if (string.IsNullOrEmpty(path)) throw new ArgumentException("path");
+
+            var req = (HttpWebRequest)WebRequest.Create(url);
+            using (var stream = req.GetResponse().GetResponseStream())
+            {
+                if (stream == null) throw new Exception("stream is null");
+                const int bufferSize = 4096;
+
+                using (var fs = File.Open(path, FileMode.Create))
+                {
+                    var buffer = new byte[bufferSize];
+                    int readed;
+                    while ((readed = stream.Read(buffer, 0, bufferSize)) != 0)
+                    {
+                        fs.Write(buffer, 0, readed);
+                    }
+                }
             }
         }
